Add docs
Refactor file structure and tests

Use ps in tests

Fix microphysics name
